Move get_param/artifact funcs into TemplateInvocatorSubNodeMixin
